add token to request status update call
