--- app/slack-calls/send-message.ts
@@ -1,14 +1,0 @@
-import { post } from 'request-promise';
-import { Message } from 'slack';
-
-export async function send(url: string, message: Message) {
-  const params = {
-    url,
-    method: 'POST',
-    simple: false,
-    body: JSON.stringify(message)
-  };
-  console.log('Replying to message with params', JSON.stringify(params));
-  const response = await post(params);
-  console.log('Received response', JSON.stringify(response));
-}
